Got simultaneous dragging of pieces and the board, depending on what you tap first. Doesn't work on the first tap for some reason (dragging a piece first works, but dragging the board first doesn't animate). Also simultaneous zooming works, however 1. it doesn't zoom towards the center of the screen (or the center of your pinch, but that seems very tricky to get right), and 2. if one of your finger drags a piece and *then* you zoom, that piece is picked up and doesn't go down for the rest of the animation.
Also, got piece animations to work, but commented them out. They kinda work, especially when picking up the piece, however 1. the piece MUST transition from scale = 0 or weird things happen (it will appear from the center of the board, not the square). 2. When placing a piece down, the piece will appear on the board below the shadow/highlight and it will appear before it has been animated to be in that position. 3. The final position of the piece/circle/square animation lags behind the actual final position of the piece, leading to a discrepancy between the animation and where the piece goes when flicking the piece